requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-google==12.0.0
google-cloud==3.0.0
pandas==2.1.4
paramiko==3.5.1
pendulum==3.0.0
protobuf==5.29.3

pysftp~=0.2.9